requirements.txt
datasets==3.6.0
openai==1.99.9
azure-identity==1.22.0
tenacity==9.1.2
pyyaml==6.0.2
pytest==8.3.5
gitignore_parser==0.1.12
termcolor==3.1.0
transformers==4.51.3
tiktoken==0.9.0
docker==7.1.0
swebench==4.0.3
swesmith==0.0.4
prompt_toolkit==3.0.51
anthropic==0.51.0
jinja2==3.1.6
rich==14.0.0
kubernetes==33.1.0
